requirements.txt
django
Pillow
django-colorfield